classes now have functionality to retrive data quickly efficiently, and packaged
