+++ server/test.js
@@ -1,5 +1,5 @@
 
-const { states } = require('./weatherUtils')
+const { states } = require('./weatherUtils/states')
