+++ server/weatherUtils/Weather/WeatherTrends.js
@@ -1,0 +1,9 @@
+
+const { WeatherInfo } = require('./WeatherInfo') // Importing the parent class
+
+
+
+
+
+
+module.exports = { WeatherTrends: WeatherTrends }
